requirements.txt
streamlit>=1.32.0
streamlit-extras>=0.3.6
firebase-admin>=6.4.0
langchain>=0.1.11
langchain_groq>=0.0.6
langchain_community>=0.0.24
langchain_core>=0.1.30
requests>=2.31.0
pandas>=2.2.0
openai>=1.12.0
python-dotenv>=1.0.1
requests-toolbelt>=1.0.0
pycryptodome>=3.20.0
uuid>=1.30.0
langchain-chroma
python-magic-bin; platform_system == "Windows"
unstructured
pdf2image
pdfminer.six
pypdf
openpyxl
beautifulsoup4
nltk
PyGithub==2.1.1
requests==2.31.0
unstructured[image]
pillow